+++ CHANGELOG.md
@@ -9,6 +9,7 @@ You can also check [on GitHub](https://github.com/nextcloud/news/releases), the 
 ### Changed
 - If title of feed is empty during creation set hostname of feed as title (#2872)
 - Add command to import OPML file
+- Add API to import OPML file or request body
 
 ### Fixed
 - Feed without Title returned by DB causes exception (#2872)